+++ .github/files/lint-project-structure.sh
@@ -165,16 +165,39 @@ for PROJECT in projects/*/*; do
 			echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::Set \`.repository.type\` to \"git\", as the monorepo is a git repository."
 		fi
 		URL="$(jq -r '.url' <<<"$JSON")"
-		if [[ "$URL" != "https://github.com/Automattic/jetpack.git" && "$URL" != "https://github.com/Automattic/jetpack" ]]; then
+
+		# Published packages need to point to the mirror repo. Unpublished packages can point to mirror or monorepo.
+		declare -A OKURLS=()
+		MIRROR=$( jq -r '.extra["mirror-repo"]' "$PROJECT/composer.json" )
+		OKURLS["git+https://github.com/$MIRROR.git"]=''
+		if jq -e '.extra["npmjs-autopublish"]' "$PROJECT/composer.json" >/dev/null; then
+			ERR1="Set \`.repository.url\` for published packages to point to the mirror repo in npm's canonical format, i.e. \"git+https://github.com/$MIRROR.git\"."
+		else
+			OKURLS["https://github.com/$MIRROR"]=''
+			OKURLS["https://github.com/$MIRROR.git"]=''
+			OKURLS["https://github.com/Automattic/jetpack"]="$PROJECT"
+			OKURLS["https://github.com/Automattic/jetpack.git"]="$PROJECT"
+			OKURLS["git+https://github.com/Automattic/jetpack.git"]="$PROJECT"
+			ERR1="Set \`.repository.url\` to point to the monorepo or mirror repo, e.g. \"https://github.com/Automattic/jetpack\" or \"git+https://github.com/$MIRROR.git\"."
+		fi
+
+		if [[ ! -v OKURLS["$URL"] ]]; then
 			EXIT=1
 			LINE=$(jq --stream -r 'if length == 1 then .[0][:-1] else .[0] end | if . == ["repository","url"] then ",line=\( input_line_number )" else empty end' "$PROJECT/package.json")
-			echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::Set \`.repository.url\` to point to the monorepo, i.e. \"https://github.com/Automattic/jetpack\"."
-		fi
-		TMP="$(jq -r '.directory' <<<"$JSON")"
-		if [[ "$TMP" != "$PROJECT" ]]; then
-			EXIT=1
-			LINE=$(jq --stream -r 'if length == 1 then .[0][:-1] else .[0] end | if . == ["repository","directory"] then ",line=\( input_line_number )" else empty end' "$PROJECT/package.json")
-			echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::Set \`.repository.directory\` to point to the project's path within the monorepo, i.e. \"$PROJECT\"."
+			echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::$ERR1"
+		elif [[ -z "${OKURLS["$URL"]}" ]]; then
+			if jq -e 'has( "directory" )' <<<"$JSON" &>/dev/null; then
+				EXIT=1
+				LINE=$(jq --stream -r 'if length == 1 then .[0][:-1] else .[0] end | if . == ["repository","directory"] then ",line=\( input_line_number )" else empty end' "$PROJECT/package.json")
+				echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::When \`.repository.url\` is set to the mirror repo, \`.repository.directory\` should not be set."
+			fi
+		else
+			TMP="$(jq -r '.directory' <<<"$JSON")"
+			if [[ "$TMP" != "${OKURLS["$URL"]}" ]]; then
+				EXIT=1
+				LINE=$(jq --stream -r 'if length == 1 then .[0][:-1] else .[0] end | if . == ["repository","directory"] then ",line=\( input_line_number )" else empty end' "$PROJECT/package.json")
+				echo "::error file=$PROJECT/package.json${LINE:-$LINE2}::Set \`.repository.directory\` to point to the project's path within the specified repo, i.e. \"${OKURLS["$URL"]}\"."
+			fi
 		fi
 	fi