+++ projects/packages/publicize/_inc/hooks/use-social-preview-post-data/types.ts
@@ -1,9 +1,7 @@
-export type PostData = {
-	title: string;
-	description: string;
-	url: string;
+import { LinkPreviewData } from '../use-link-preview-post-data/types';
+
+export type PostPreviewData = LinkPreviewData & {
 	excerpt: string;
-	image: string;
 	media: Array< {
 		type: string;
 		url: string;