+++ projects/js-packages/number-formatters/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.17] - 2026-02-12
+### Changed
+- Update package dependencies. [#47099]
+
 ## [1.0.16] - 2026-01-14
 ### Changed
 - Internal updates.
@@ -97,6 +101,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Initial release
 - Basic number formatting functionality
 
+[1.0.17]: https://github.com/Automattic/number-formatters/compare/1.0.16...1.0.17
 [1.0.16]: https://github.com/Automattic/number-formatters/compare/1.0.15...1.0.16
 [1.0.15]: https://github.com/Automattic/number-formatters/compare/1.0.14...1.0.15
 [1.0.14]: https://github.com/Automattic/number-formatters/compare/1.0.13...1.0.14