+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.7.24 - 2026-02-16
+### Changed
+- Update package dependencies. [#47130]
+
+## 3.7.23 - 2026-02-12
+### Changed
+- Added WebP to default file extensions handled by FileRule. [#47082]
+- Update package dependencies. [#47095]
+- Update package dependencies. [#47099]
+- Update package dependencies. [#47101]
+
 ## 3.7.22 - 2026-02-10
 ### Changed
 - Update package dependencies. [#47002]