+++ projects/packages/post-list/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.6] - 2026-02-16
+### Changed
+- Update dependencies. [#47109]
+
 ## [0.9.5] - 2026-02-10
 ### Changed
 - Update package dependencies. [#47002]
@@ -310,6 +314,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated the default columns displayed on the post and page list screens
 - Refactored thumbnail preview to function server side. All javascript removed.
 
+[0.9.6]: https://github.com/automattic/jetpack-post-list/compare/v0.9.5...v0.9.6
 [0.9.5]: https://github.com/automattic/jetpack-post-list/compare/v0.9.4...v0.9.5
 [0.9.4]: https://github.com/automattic/jetpack-post-list/compare/v0.9.3...v0.9.4
 [0.9.3]: https://github.com/automattic/jetpack-post-list/compare/v0.9.2...v0.9.3